+++ python/paddle/distributed/fleet/meta_optimizers/common.py
@@ -99,8 +99,8 @@ class CollectiveHelper(object):
                 OP_ROLE_KEY: OpRole.Forward
             })
 
-    def _wait(self, current_endpoint, endpoints, wait_port):
-        assert (wait_port)
+    def _wait(self, current_endpoint, endpoints):
+        assert (self.wait_port)
         other_endpoints = endpoints[:]
         other_endpoints.remove(current_endpoint)
         wait_server_ready(other_endpoints)
